Fix [General] Chips pop-up is hidden for the first table item
The pop-up for the hidden chips was moved to the 'overlay_container' to prevent other elements from overlapping the pop-up.
