--- src/elements/HiddenChipsBlock/hiddenChipsBlock.scss
@@ -1,58 +1,0 @@
-@import '~igz-controls/scss/variables';
-@import '~igz-controls/scss/colors';
-@import '~igz-controls/scss/shadows';
-@import '~igz-controls/scss/borders';
-@import '~igz-controls/scss/mixins';
-
-.chip-block-hidden {
-  position: absolute;
-  right: 0;
-  z-index: 6;
-  display: flex;
-  flex-flow: row wrap;
-  min-width: 250px;
-  max-width: 45vw;
-  padding: 10px;
-  background-color: $white;
-  border-radius: $mainBorderRadius;
-  box-shadow: $hiddenChipsBlockShadow;
-  opacity: 0;
-
-  &_bottom {
-    bottom: 112%;
-
-    &::after {
-      @include hiddenChipBlockArrow();
-
-      bottom: -6px;
-    }
-  }
-
-  &_top {
-    top: 112%;
-
-    &::before {
-      @include hiddenChipBlockArrow();
-
-      top: -6px;
-    }
-  }
-
-  &_left {
-    &::before {
-      right: 20px;
-    }
-
-    &::after {
-      right: 20px;
-    }
-  }
-
-  &_right {
-    right: unset;
-  }
-
-  &_visible {
-    opacity: 1;
-  }
-}
